cell/pyproject.toml
[project]
name = "cell"
version = "0.1.0"
description = "Single-cell multi-agent cross-verification runtime"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.6",
    "httpx>=0.27",
    "temporalio>=1.7",
    "pyyaml>=6.0",
    "RestrictedPython>=7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-timeout>=2.2",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["cell*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

